Worked on models,services, daos

+++ shop/shop-base/src/main/java/com/example/shop/base/dao/CategoryDao.java
@@ -11,7 +11,7 @@ package com.example.shop.base.dao;
 import com.example.shop.base.model.Category;
 
 
-public interface CategoryDao extends GenericDao<Category, Long>
+public interface CategoryDao extends GenericDao<Category, Integer>
 {
     Category findByName(String categoryName);
 }
