+++ shop/shop-base/src/main/java/com/example/shop/base/dao/GenericDao.java
@@ -29,6 +29,6 @@ public interface GenericDao<E, PK extends Serializable>
     E findById(PK id);
 
 
-    long size();
+    Integer size();
 
 }
