+++ shop/shop-base/src/main/java/com/example/shop/base/dao/ProductDao.java
@@ -11,7 +11,7 @@ package com.example.shop.base.dao;
 import com.example.shop.base.model.Product;
 
 
-public interface ProductDao extends GenericDao<Product, Long>
+public interface ProductDao extends GenericDao<Product, Integer>
 {
     Product findByName(String productName);
 }
