+++ shop/shop-base/src/main/java/com/example/shop/base/service/OrderService.java
@@ -15,7 +15,7 @@ import javax.jws.WebService;
 
 
 @WebService
-public interface OrderService extends GenericService<Order, Long>
+public interface OrderService extends GenericService<Order, Integer>
 {
     void setOrderDao(OrderDao orderDao);
 }
