+++ shop/shop-base/src/main/java/com/example/shop/base/service/impl/CategoryServiceImpl.java
@@ -28,7 +28,7 @@ public class CategoryServiceImpl implements CategoryService
 
 
     @Override
-    public Category getById(Long id)
+    public Category getById(Integer id)
     {
         return this.categoryDao.findById(id);
     }
