+++ shop/shop-base/src/main/java/com/example/shop/base/service/impl/OrderServiceImpl.java
@@ -29,7 +29,7 @@ public class OrderServiceImpl implements OrderService
 
 
     @Override
-    public Order getById(Long id)
+    public Order getById(Integer id)
     {
         final Order order = this.orderDao.findById(id);
         return order;
