+++ shop/shop-base/src/main/java/com/example/shop/base/service/impl/ProductServiceImpl.java
@@ -29,7 +29,7 @@ public class ProductServiceImpl implements ProductService
 
 
     @Override
-    public Product getById(Long id)
+    public Product getById(Integer id)
     {
         final Product product = this.productDao.findById(id);
         return product;
