+++ shop/shop-base/src/main/java/com/example/shop/base/service/impl/UserServiceImpl.java
@@ -30,7 +30,7 @@ public class UserServiceImpl implements UserService
 
 
     @Override
-    public User getById(Long id)
+    public User getById(Integer id)
     {
         final User user = this.userDao.findById(id);
         return user;
